Move schema options to new test file
